pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lolz-bump"
version = "0.1.0"
description = "LOLZ forum thread bump scheduler"
requires-python = ">=3.11"
dependencies = [
  "aiohttp>=3.10,<4.0",
  "APScheduler>=3.10,<4.0",
  "SQLAlchemy>=2.0,<3.0",
  "pydantic>=2.7,<3.0",
  "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-asyncio>=1.0,<2.0",
]

[tool.setuptools.packages.find]
include = ["lolz_bump*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
